+++ src/TPP/dataloader/lognormmix/log_dataset.py
@@ -69,7 +69,7 @@ class LogNormDataset(utils.data.Dataset):
         self.time_seq = [np.array(seq, dtype = np.float32) for seq in self.time_seq]
         self.score = [np.array(seq, dtype = np.float32) for seq in self.score]
         self.intensity = [np.array(seq, dtype = np.float32) for seq in self.intensity]
-        self.event = [np.array(seq, dtype = np.int32) for seq in self.event]
+        self.event = [np.array(seq, dtype = np.int64) for seq in self.event]
 
 
     def __getitem__(self, index):
